Forgot To Commit so Large Commit Here
Implemented Mathias' Touch Script, and extended for computer debugging (since a touch does not activate on mouse click)

Made a Deconstructable object which on click instantiates a random amount of "numbers", it also randomly chooses what numbers to instantiate.

Made a set of simple number prefabs.

Made a movement functionality based on Unitys pathfinding algorithm.

Hugo will move on touch, and on mouse click. (depending on whether debug is true in the TouchHandler)

Also added a Tag
- Touchable (objects with the Touchable script, who are supposed to be tapped)

Also added a few Layers
- Numbers (for the pieces which are dropped on deconstruction)
- Deconstructable
- Player

These layers are for controlling which objects are supposed to collide with which. For example, the numbers do not collide with the deconstructables, neither do the collide with hugo, since this would in some cases push hugo.